requirements.txt
Flask
Flask-Babel
Flask-SQLAlchemy
numpy
pillow
tensorflow
